requirements.txt
python-dotenv==0.8.2
mysql-connector-python>=8.0,<8.1